requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
requests==2.31.0
google-generativeai==0.3.2
beautifulsoup4==4.12.2
tqdm==4.66.1
jinja2==3.1.2
pydantic==2.5.0
python-dotenv==1.0.0